Refactor into separate modules
