+++ packages/gatsby/src/internal-plugins/partytown/gatsby-browser.tsx
@@ -1,35 +1,6 @@
 import { collectedScriptsByPage } from "gatsby-script"
-import { partytownSnippet } from "@builder.io/partytown/integration"
+import { injectPartytownSnippet } from "./utils/inject-partytown-snippet"
 import type { GatsbyBrowser } from "gatsby"
-import type { ScriptProps } from "gatsby-script"
-
-function getForwards(collectedScripts: Array<ScriptProps>): Array<string> {
-  const forwards = collectedScripts?.flatMap(
-    (script: ScriptProps) => script?.forward || []
-  )
-  return forwards
-}
-
-function injectPartytownSnippet(pathname: string): void {
-  const collectedScripts = collectedScriptsByPage.get(pathname)
-
-  if (!collectedScripts.length) {
-    return
-  }
-
-  const existingSnippet = document.querySelector(`[data-partytown]`)
-
-  if (existingSnippet) {
-    existingSnippet.remove()
-  }
-
-  const forwards = getForwards(collectedScripts)
-
-  const snippet = document.createElement(`script`)
-  snippet.dataset.partytown = ``
-  snippet.innerHTML = partytownSnippet({ forward: forwards })
-  document.head.appendChild(snippet)
-}
 
 export const onInitialClientRender: GatsbyBrowser[`onInitialClientRender`] =
   (): void => {
