+++ packages/gatsby/src/internal-plugins/partytown/utils/get-forwards.ts
@@ -1,0 +1,9 @@
+import type { ScriptProps } from "gatsby-script"
+
+export function getForwards(
+  collectedScripts: Array<ScriptProps>
+): Array<string> {
+  return collectedScripts?.flatMap(
+    (script: ScriptProps) => script?.forward || []
+  )
+}
